+++ components/CardsReference.vue
@@ -49,7 +49,8 @@ const TokyoFeverConsultationCenterReportsNumberCard = () =>
     '@/components/index/CardsReference/TokyoFeverConsultationCenterReportsNumber/Card.vue'
   )
 // 都営地下鉄の利用者数の推移
-const MetroCard = () => import('@/components/cards/MetroCard.vue')
+const MetroCard = () =>
+  import('@/components/index/CardsReference/Metro/Card.vue')
 // 都庁来庁者数の推移
 const AgencyCard = () => import('@/components/cards/AgencyCard.vue')
 /* eslint-enable simple-import-sort/imports */
